# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PerioVision AI API",
    "version": "2.0",
    "description": "Research prototype for periodontal decision support. Every JSON response uses the envelope {data, meta, error, mode}. Authenticate with POST /api/auth/login, then send 'Authorization: Bearer <access_token>'. x-permission names the RBAC permission each operation requires (see docs/SECURITY.md)."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Envelope": {
        "type": "object",
        "required": [
          "data",
          "meta",
          "error",
          "mode"
        ],
        "properties": {
          "data": {
            "description": "Payload (object, list or null on error)"
          },
          "meta": {
            "type": "object",
            "description": "Extra information such as counts"
          },
          "error": {
            "nullable": true,
            "type": "object",
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ]
          }
        }
      },
...
      "PatientIn": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "age": {
            "anyOf": [
              {
                "maximum": 120,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Age"
          },
          "sex": {
            "anyOf": [
              {
                "enum": [
                  "female",
                  "male",
                  "other",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Sex"
          },
          "contact": {
            "anyOf": [
              {
                "maxLength": 60,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Contact"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Notes"
          },
          "smoking_status": {
            "anyOf": [
              {
                "enum": [
                  "never",
                  "former",
                  "current",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "unknown",
            "title": "Smoking Status"
          },
          "cigarettes_per_day": {
            "anyOf": [
              {
                "maximum": 100,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Cigarettes Per Day"
          },
          "diabetic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Diabetic"
          },
          "hba1c": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 3.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Hba1C"
          },
          "teeth_lost_perio": {
            "anyOf": [
              {
                "maximum": 32,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Teeth Lost Perio"
          }
        },
        "required": [
          "name"
        ],
        "title": "PatientIn",
        "type": "object"
      },
      "PatientUpdateIn": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Name"
          },
          "age": {
            "anyOf": [
              {
                "maximum": 120,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Age"
          },
          "sex": {
            "anyOf": [
              {
                "enum": [
                  "female",
                  "male",
                  "other",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Sex"
          },
          "contact": {
            "anyOf": [
              {
                "maxLength": 60,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Contact"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Notes"
          },
          "smoking_status": {
            "anyOf": [
              {
                "enum": [
                  "never",
                  "former",
                  "current",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Smoking Status"
          },
          "cigarettes_per_day": {
            "anyOf": [
              {
                "maximum": 100,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Cigarettes Per Day"
          },
          "diabetic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Diabetic"
          },
          "hba1c": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 3.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Hba1C"
          },
          "teeth_lost_perio": {
            "anyOf": [
              {
                "maximum": 32,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Teeth Lost Perio"
          }
        },
        "title": "PatientUpdateIn",
        "type": "object"
      },
...
    }
  },
  "paths": {
...
    "/api/patients": {
      "get": {
        "summary": "List or search patients (exact name or pseudonym)",
        "operationId": "patients_list_patients_get",
        "tags": [
          "patients"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "patient:read"
      },
      "post": {
        "summary": "Create a patient (identity fields are encrypted)",
        "operationId": "patients_create_patient_post",
        "tags": [
          "patients"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          },
          "422": {
            "description": "Validation failed; error.details lists each field"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "patient:write",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientIn"
              }
            }
          }
        }
      }
    },
    "/api/patients/{patient_id}": {
      "get": {
        "summary": "Patient detail with visit history",
        "operationId": "patients_get_patient_get",
        "tags": [
          "patients"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "patient:read",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "patch": {
        "summary": "Update patient details or clinical risk factors",
        "operationId": "patients_update_patient_patch",
        "tags": [
          "patients"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          },
          "422": {
            "description": "Validation failed; error.details lists each field"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "patient:write",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdateIn"
              }
            }
          }
        }
      }
    },
...
    "/api/patients/{patient_id}/care-team": {
      "post": {
        "summary": "Add a clinician to a patient's care team",
        "operationId": "patients_add_care_team_post",
        "tags": [
          "patients"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "admin:users",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
  }
}